refactor: style 코드 모듈로 분리
