+++ src/views/Payment/components/CouponMessage/CouponMessage.style.ts
@@ -1,0 +1,8 @@
+import {styled} from "styled-components";
+
+export const Message = styled.span`
+  padding-left: 1rem;
+  font-size: 1.6rem;
+  font-weight: 800;
+  line-height: 20px;
+`;
