+++ src/views/Payment/components/OrderList/OrderList.style.ts
@@ -1,0 +1,6 @@
+import {styled} from "styled-components";
+
+export const WrapperPage = styled.main`
+  display: flex;
+  flex-direction: column;
+`;
