+++ src/views/Payment/components/OrderList/OrderList.tsx
@@ -2,22 +2,20 @@ import { styled } from 'styled-components';
 
 import { useOrderList } from '@views/Payment/recoil/orderListState';
 import { OrderItem } from '../Order';
+import * as S from './OrderList.style'
 
 function OrderList() {
   const orderList = useOrderList();
 
+  // TODO: key
   return (
-    <WrapperPage>
+    <S.WrapperPage>
       {[...orderList].reverse().map((order) => (
         <OrderItem order={order} hasDetail />
       ))}
-    </WrapperPage>
+    </S.WrapperPage>
   );
 }
 
 export default OrderList;
 
-const WrapperPage = styled.main`
-  display: flex;
-  flex-direction: column;
-`;
