+++ src/views/Payment/components/OrderListButton/OrderListButton.style.ts
@@ -1,0 +1,8 @@
+import {styled} from "styled-components";
+
+export const HeaderIconWrapper = styled.button`
+  color: ${({ theme }) => theme.lightColor};
+  font-size: 24px;
+  display: flex;
+  align-items: center;
+`;
