+++ src/views/Payment/components/OrderListButton/index.ts
@@ -1,0 +1,1 @@
+export { default as OrderListButton} from './OrderListButton'
